requirements.txt
transformers>=4.35.0
torch>=2.2.0
pillow>=10.0.0
gradio==3.50.2
accelerate>=0.25.0
